Include "ppport.h" in the tutorials; change "" to <>
There are some problems with compiling this on Windows and other
systems:

http://matrix.cpantesters.org/?dist=XS-Tutorial+0.04

I haven't inspected every case, but judging from the ones I have
looked at, the lack of #include ppport.h in the .xs files might be the
reason.

Also, stdint.h is a system file so it should be <stdint.h>, although
"stdint.h" will work.